Drop special support for point spans at the end of files (#31)
Drop special support for point spans at the end of files

This broke point spans at the beginning of lines elsewhere in the
file. Unfortunately, there's no way to disambiguate these currently.
SourceSpanWithContext doesn't have any information about where the
span appears in the context, so for point spans at column 0, it could
be at the beginning of the line or after the trailing newline.

We should eventually add something like
SourceSpanWithContext.indexInContext, but that'll require a breaking
release, since downstream users implement FileSpan as an interface.